refactor(agents): separate prepared runtime retention
Move existing in-memory retirement and retention logic into its owner
module and place synchronous auth invalidation beside auth publication.
Preserve admission and lease guards, identity checks and refresh ordering.

Related: #140674
