+++ src/agents/prepared-model-runtime-auth-publication.ts
@@ -293,3 +293,35 @@ export class PreparedModelRuntimeAuthPublicationOwner {
     return rejected;
   }
 }
+
+export function invalidatePreparedModelRuntimeOwnersForAuthMutation(
+  owners: Map<string, PreparedModelRuntimeOwner>,
+  normalizedEvent: PreparedModelRuntimeAuthMutation,
+): {
+  invalidatedOwners: PreparedModelRuntimeOwner[];
+  invalidatedConfiguredAgentIds: Set<string>;
+} {
+  const staleError = new Error("prepared model runtime owner is stale after auth mutation");
+  const invalidatedOwners: PreparedModelRuntimeOwner[] = [];
+  const invalidatedConfiguredAgentIds = new Set<string>();
+  for (const owner of owners.values()) {
+    if (
+      !normalizedEvent.affectsInheritedStores &&
+      owner.input.agentDir !== normalizedEvent.agentDir &&
+      owner.input.inheritedAuthDir !== normalizedEvent.agentDir
+    ) {
+      continue;
+    }
+    invalidatedOwners.push(owner);
+    owner.generation += 1;
+    owner.needsRefresh = true;
+    owner.refreshError = staleError;
+    if (normalizedEvent.profileSetChanged) {
+      owner.catalogStale = true;
+    }
+    if (owner.provenance === "configured" && owner.input.agentId) {
+      invalidatedConfiguredAgentIds.add(owner.input.agentId);
+    }
+  }
+  return { invalidatedOwners, invalidatedConfiguredAgentIds };
+}
